backend/pyproject.toml
[project]
name = "Rivora-backend"
version = "0.1.0"
description = "Backend of multi e vendor e commerce application"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email,timezone]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
